add dvc for dat mngeent and  versionning
