+++ .dvc/.gitignore
@@ -1,0 +1,5 @@
+/config.local
+/tmp
+/tmp
+/cache
+/cache
