+++ .dvc/config
@@ -1,0 +1,7 @@
+[core]
+    remote = minio
+['remote "minio"']
+    url = s3://dvcstore
+    endpointurl = http://172.18.0.2:9000
+    access_key_id = root
+    secret_access_key = toortoor
