+++ bash/2-run-mlops-servers.sh
@@ -8,7 +8,8 @@ project_name="${PWD##*/}"
 # run the mlops servers
 mlops_dir=../MLOPs-template/
 cd $mlops_dir
-bash bash/open-servers-browser.sh
+bash bash/run-servers.sh
+# bash bash/open-servers-browser.sh
 
 # create an alias to the jenkins outputs
 ln -s "$(pwd)/jenkins-data/jobs" ../../jenkins-outputs
@@ -24,13 +25,9 @@ cd $dir
 echo && echo "[${PROJECT_NAME}][Docker-Compose] Updating the configuration file..."
 
 if [ "$MLFLOW_SERVER_URL" != "" ] ; then
-	#### --------------------   UPDATE THE MLFLOW_URI  -------------------------
-	echo "   -> updating the mlflow_uri in <config/config.yaml> file "
-	sed -i "s|mlflow_uri:.*|mlflow_uri: $MLFLOW_SERVER_URL|g" config/config.yaml
-	echo "   -> S3_LOCAL_DIR=${S3_LOCAL_DIR}"
-	# xdg-open "${MLFLOW_SERVER_URL}"
+    #### --------------------   UPDATE THE MLFLOW_URI  -------------------------
+    echo "   -> updating the mlflow_uri in <config/config.yaml> file "
+    sed -i "s|mlflow_uri:.*|mlflow_uri: $MLFLOW_SERVER_URL|g" config/config.yaml
+    echo "   -> S3_LOCAL_DIR=${S3_LOCAL_DIR}"
+    # xdg-open "${MLFLOW_SERVER_URL}"
 fi
-
-# run the projects servers
-echo $(pwd)
-bash bash/4-open-app-servres-in-browser.sh
