+++ dvc/0-setup.sh
@@ -1,0 +1,2 @@
+# install DVC package
+pip install dvc  dvc_s3 dvc[all]
