--- src/utils/data_preparation.py
@@ -1,5 +1,0 @@
-# from transformers import AutoTokenizer
-
-# # Tokenize text using Hugging Face’s tokenizer
-# tokenizer = AutoTokenizer.from_pretrained("distilbert-base-uncased")
-# tokens = tokenizer("This is a test sentence", return_tensors="pt")
